requirements.txt
streamlit==1.24.0
pandas==2.1.0
sqlalchemy==2.0.9
cryptography==39.0.1
PyPDF2==3.0.1
python-docx==0.8.11
